# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Kasi Dash API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /careers/{jobId}:
    get:
      operationId: getJobListing
      tags: [careers]
      summary: Get job listing by ID
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Job listing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobListing"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /apply/driver:
    post:
      operationId: submitDriverApplication
      tags: [applications]
      summary: Submit driver application
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DriverApplicationBody"
      responses:
        "201":
          description: Submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApplicationResponse"
  /apply/vendor:
    post:
      operationId: submitVendorApplication
      tags: [applications]
      summary: Submit vendor/partner application
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VendorApplicationBody"
      responses:
        "201":
          description: Submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApplicationResponse"
  /apply/job:
    post:
      operationId: submitJobApplication
      tags: [applications]
      summary: Submit job application
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobApplicationBody"
      responses:
        "201":
          description: Submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApplicationResponse"
...
  /staff/applications/drivers/{id}:
    get:
      operationId: getDriverApplication
      tags: [staff]
      summary: Get driver application
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Application
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DriverApplication"
    patch:
      operationId: updateDriverApplication
      tags: [staff]
      summary: Update driver application status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateApplicationStatusBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DriverApplication"
...
  /staff/applications/vendors/{id}:
    patch:
      operationId: updateVendorApplication
      tags: [staff]
      summary: Update vendor application status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateApplicationStatusBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VendorApplication"
...
  /staff/applications/jobs/{id}:
    patch:
      operationId: updateJobApplication
      tags: [staff]
      summary: Update job application status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateApplicationStatusBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobApplication"
  /staff/careers:
    get:
      operationId: staffListJobListings
      tags: [staff]
      summary: List all job listings (including inactive)
      responses:
        "200":
          description: All listings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/JobListing"
    post:
      operationId: createJobListing
      tags: [staff]
      summary: Create a job listing
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateJobListingBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobListing"
  /staff/careers/{id}:
    patch:
      operationId: updateJobListing
      tags: [staff]
      summary: Update a job listing
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateJobListingBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobListing"
    delete:
      operationId: deleteJobListing
      tags: [staff]
      summary: Delete a job listing
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiSuccess"
components:
  schemas:
...
    ApiError:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error, message]
    ApiSuccess:
      type: object
      properties:
        success:
          type: boolean
      required: [success]
...
    JobListing:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        department:
          type: string
        location:
          type: string
        type:
          type: string
          enum: [full_time, part_time, contract, internship]
        description:
          type: string
        requirements:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, title, department, location, type, description, requirements, isActive, createdAt, updatedAt]
    CreateJobListingBody:
      type: object
      properties:
        title:
          type: string
        department:
          type: string
        location:
          type: string
        type:
          type: string
          enum: [full_time, part_time, contract, internship]
        description:
          type: string
        requirements:
          type: string
        isActive:
          type: boolean
      required: [title, department, location, type, description, requirements]
    UpdateJobListingBody:
      type: object
      properties:
        title:
          type: string
        department:
          type: string
        location:
          type: string
        type:
          type: string
          enum: [full_time, part_time, contract, internship]
        description:
          type: string
        requirements:
          type: string
        isActive:
          type: boolean
    DriverApplicationBody:
      type: object
      properties:
        fullName:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        idNumber:
          type: string
        vehicleType:
          type: string
          enum: [bicycle, motorcycle, car, bakkie, van]
        vehicleRegistration:
          type: string
        licenseNumber:
          type: string
        township:
          type: string
        message:
          type: string
      required: [fullName, email, phone, idNumber, vehicleType, licenseNumber, township]
    VendorApplicationBody:
      type: object
      properties:
        businessName:
          type: string
        contactName:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        businessType:
          type: string
        township:
          type: string
        description:
          type: string
        website:
          type: string
      required: [businessName, contactName, email, phone, businessType, township]
    JobApplicationBody:
      type: object
      properties:
        jobListingId:
          type: string
        fullName:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        coverLetter:
          type: string
        linkedIn:
          type: string
        experience:
          type: string
      required: [jobListingId, fullName, email, phone, coverLetter]
    ApplicationResponse:
      type: object
      properties:
        id:
          type: string
        status:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, status, createdAt]
    UpdateApplicationStatusBody:
      type: object
      properties:
        status:
          type: string
          enum: [pending, reviewing, approved, rejected]
        notes:
          type: string
      required: [status]
    DriverApplication:
      type: object
      properties:
        id:
          type: string
        fullName:
          type: string
        email:
          type: string
        phone:
          type: string
        idNumber:
          type: string
        vehicleType:
          type: string
        vehicleRegistration:
          type: string
        licenseNumber:
          type: string
        township:
          type: string
        message:
          type: string
        status:
          type: string
          enum: [pending, reviewing, approved, rejected]
        notes:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, fullName, email, phone, vehicleType, licenseNumber, township, status, createdAt, updatedAt]
    VendorApplication:
      type: object
      properties:
        id:
          type: string
        businessName:
          type: string
        contactName:
          type: string
        email:
          type: string
        phone:
          type: string
        businessType:
          type: string
        township:
          type: string
        description:
          type: string
        website:
          type: string
        status:
          type: string
          enum: [pending, reviewing, approved, rejected]
        notes:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, businessName, contactName, email, phone, businessType, township, status, createdAt, updatedAt]
    JobApplication:
      type: object
      properties:
        id:
          type: string
        jobListingId:
          type: string
        jobTitle:
          type: string
        fullName:
          type: string
        email:
          type: string
        phone:
          type: string
        coverLetter:
          type: string
        linkedIn:
          type: string
        experience:
          type: string
        status:
          type: string
          enum: [pending, reviewing, approved, rejected]
        notes:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, jobListingId, fullName, email, phone, coverLetter, status, createdAt, updatedAt]
...